Abstract native helper tools for OSX
