+++ netfox/NFXAssets.swift
@@ -8,7 +8,7 @@
 import Foundation
 import UIKit
 
-enum NFXImage {
+enum NFXAssetName {
     case SETTINGS
     case INFO
     case STATISTICS
@@ -16,7 +16,7 @@ enum NFXImage {
 
 class NFXAssets
 {
-    class func getImage(image: NFXImage) -> NSData
+    class func getImage(image: NFXAssetName) -> NSData
     {
         var base64Image: String
